+++ src/Laplace/lfmm3dwrap_legacy.f
@@ -156,7 +156,7 @@ c                  3 => tolerance =.5d-9
 c                  4 => tolerance =.5d-12
 c                  5 => tolerance =.5d-15
 c
-c       nsource: integer:  number of sources
+c       nsource: integer(8):  number of sources
 c       source: real *8 (3,nsource):  source locations
 c       ifcharge:  charge computation flag
 c                  ifcharge = 1   =>  include charge contribution
@@ -170,7 +170,7 @@ c       dipvec: real *8 (3,nsource): dipole orientation vectors.
 c
 c       ifpot:  potential flag (1=compute potential, otherwise no)
 c       iffld:  field flag (1=compute field, otherwise no)
-c       ntarget: integer:  number of targets
+c       ntarget: integer(8):  number of targets
 c       target: real *8 (3,ntarget):  target locations
 c       ifpottarg:  target potential flag 
 c                   (1=compute potential, otherwise no)
@@ -186,19 +186,19 @@ c       fld: complex *16 (3,nsource): field (-gradient) at source locations
 c       pottarg: complex *16 (ntarget): potential at target locations 
 c       fldtarg: complex *16 (3,ntarget): field (-gradient) at target locations 
 c
-      integer ier,iprec,nsource
-      integer ifcharge,ifdipole,iper
+      integer(8) ier,iprec,nsource
+      integer(8) ifcharge,ifdipole,iper
       double precision source(3,nsource)
       
       double complex charge(*),dipstr(*)
       double precision dipvec(3,*)
 
-      integer ifpot,iffld,ifpottarg,iffldtarg
+      integer(8) ifpot,iffld,ifpottarg,iffldtarg
       double complex  pot(*),fld(3,*)
       double complex  pottarg(*),fldtarg(3,*)
 
-      integer nd,ifpgh,ifpghtarg
-      integer ntarg
+      integer(8) nd,ifpgh,ifpghtarg
+      integer(8) ntarg
       double precision targ(3,*)
       double complex, allocatable :: dipvec_in(:,:)
       
@@ -208,7 +208,7 @@ c
       double complex hess(6),hesstarg(6)
       double precision eps
 
-      integer i
+      integer(8) i
      
 c     set fmm tolerance based on iprec flag.
 c
@@ -349,7 +349,7 @@ c       without the (1/4 pi) scaling.  Self-interactions are not-included.
 c   
 c       INPUT PARAMETERS:
 c
-c       nsource: integer:  number of sources
+c       nsource: integer(8):  number of sources
 c       source: real *8 (3,nsource):  source locations
 c       ifcharge:  charge computation flag
 c                  ifcharge = 1   =>  include charge contribution
@@ -363,7 +363,7 @@ c       dipvec: real *8 (3,nsource): dipole orientation vectors.
 c
 c       ifpot:  potential flag (1=compute potential, otherwise no)
 c       iffld:  field flag (1=compute field, otherwise no)
-c       ntarg: integer:  number of targets
+c       ntarg: integer(8):  number of targets
 c       targ: real *8 (3,ntarg):  target locations
 c       ifpottarg:  target potential flag 
 c                   (1=compute potential, otherwise no)
@@ -377,9 +377,9 @@ c       fld: complex *16 (3,nsource): field (-gradient) at source locations
 c       pottarg: complex *16 (ntarg): potential at target locations 
 c       fldtarg: complex *16 (3,ntarg): field (-gradient) at target locations 
 c
-      integer nsource,ifcharge,ifdipole,ifpot,iffld,ntarg
-      integer ifpottarg,iffldtarg
-      integer nt,ns
+      integer(8) nsource,ifcharge,ifdipole,ifpot,iffld,ntarg
+      integer(8) ifpottarg,iffldtarg
+      integer(8) nt,ns
       double precision source(3,*), targ(3,*)
       double complex charge(*),dipstr(*)
       double precision, allocatable :: charge_in(:,:)
@@ -393,8 +393,8 @@ c
       double precision, allocatable :: dipvec_in(:,:,:)
       double precision thresh
 
-      integer i,j,nd
-      integer ifpgh,ifpghtarg
+      integer(8) i,j,nd
+      integer(8) ifpgh,ifpghtarg
 
       double precision xmin,xmax,ymin,ymax,zmin,zmax
       double precision bsize,btmp,sizex,sizey,sizez
@@ -402,7 +402,7 @@ c
       double complex ima
       data ima/(0.0d0,1.0d0)/
 
-      integer ntmp
+      integer(8) ntmp
 
       nt = ntarg
       ns = nsource
